Refactor absorbing boundary base classes
Refactor absorbing boundary base classes
