+++ spyro/abc/abc_layer.py
@@ -1,18 +1,11 @@
-from firedrake import Function, VTKFile
-from numpy import abs, array, ceil, inf, log10, minimum
-from sympy import divisors
-from .eik_min import Minimum_Eikonal
+from numpy import inf
+from .abc import AbsorbingBC
 from .nrbc import NRBC
-from ..solvers.modal.modal_sol import Modal_Solver
-from ..tools.error_measure import MeasureError
-from ..domains.space import create_function_space
 from ..io.basicio import parallel_print as pprint
 from .lay_len import calc_size_lay
 from ..plots.plots_habc import plot_function_layer_size
 from ..tools.abc_labeling_cases import formatting_abc_layer_type, path_to_save_abc_case
-from ..tools.habc_tools import clipping_coordinates_lay_field, extend_scalar_field_profile
-from ..utils.error_management import (validate_enum, validate_data_structure,
-                                      validate_numeric, validate_parameter)
+from ..utils.error_management import validate_enum, validate_numeric, validate_parameter
 from ..utils.freq_tools import freq_response
 from ..utils.typing import (AbsorbingBCsType, BoundaryConditionsType, HyperLayerDegreeType,
                             LayerShapeType, LayerSizeRefFrequency, NRBCBoundaryType)
@@ -27,7 +20,7 @@ from ..utils.typing import (AbsorbingBCsType, BoundaryConditionsType, HyperLayer
 # TODO: Add reference
 
 
-class ABCLayer(NRBC, MeasureError):
+class ABCLayer(AbsorbingBC):
     """Class for ABCs based on absorbing layers.
 
     Attributes
@@ -60,60 +53,22 @@ class ABCLayer(NRBC, MeasureError):
         hypershape layers ("REC", "HNX.Y" with X.Y as the hypershape degree with one
         decimal place precision) and the reference frequency ('SOU' or 'BND').
         Examples: "REC_SOU", "REC_BND", "HN2.4_SOU" or "HN2.4_BND".
-    comm : `object`
-        An object representing the communication interface for parallel processing.
-        Default is `None`.
     crit_source : `tuple`
        Critical source coordinates.
     d_norm : `float`
         Normalized element size (lmin / pad_len)
-    dimension : `int`, optional
-        Model dimension (2D or 3D). Default is 2D.
-    domain_dim : `tuple`
-        Original domain dimensions: (length_z, length_x) for 2D
-        or (length_z, length_x, length_y) for 3D.
-    dt : `float` or `None`
-        Time step used in the simulation. It is `None` if the response is not 'transient'.
-    eik_bnd : `list`
-        Properties on boundaries according to minimum values of Eikonal.
-        Structure sublist: [pnt_crit, c_bnd, eikmin, z_par, lref, sou_crit]
-        - pnt_crit : `array`
-            Critical point coordinates.
-        - c_bnd :  `float`
-            Propagation speed at critical point.
-        - eikmin : `float`
-            Minimum eikonal value in seconds.
-        - z_par :  `float`
-            Inverse of minimum Eikonal (Equivalent to c_bound/lref).
-        - lref : `float`
-            Distance to the closest source from critical point.
-        - sou_crit : `tuple`
-            Critical source coordinates.
     ele_pad : `int`
         Number of elements in the layer of edge length equal to 'lmin'.
     factor_length_pad : `float`
         Size parameter of the absorbing layer.
-    frequency: `float`
-        Frequency of the source.
-    freq_Nyquist : `float`
-        Nyquist frequency according to the time step. freq_Nyquist = 1 / (2 * dt).
     freq_ref : `float`
         Reference frequency of the wave at the critical poin on boundary.
-    func_space_type, `str`, optional
-        Type of function space for the state variable.
-        Options: 'scalar' or 'vector'. Default is None.
     FLpossible : `list`
         Possible size parameters for the absorbing layer without rounding.
     layer_geometry : `rec_lay.RectangLayer` or `hyp_lay.HyperLayer`
         An object representing the geometry of the absorbing layer.
         Options: `rec_lay.RectangLayer` for rectangular layers or `hyp_lay.HyperLayer`
         for hypershape layers.
-    length_xabc : `float`
-        Length of the domain in the x-direction with absorbing layer.
-    length_yabc : `float`
-        Length of the domain in the y-direction with absorbing layer (3D).
-    length_zabc : `float`
-        Length of the domain in the z-direction with absorbing layer.
     lmin : `float`
         Minimum mesh size.
     lref : `float`
@@ -122,42 +77,17 @@ class ABCLayer(NRBC, MeasureError):
         Path to save data for the current case study of ABCs based on absorbing layers.
     path_save : `string`
         Path to save data.
-    quadrilateral : `bool`, optional
-        Flag to indicate whether to use quadrilateral/hexahedral elements.
-        Default is `False` (triangular/tetrahedral elements).
 
     Methods
     -------
     _define_layer_shape()
         Define the shape of the absorbing layer.
-    abc_domain_dimensions()
-        Return the new dimensions of the domain with absorbing layer as a tuple.
-    abc_new_geometry()
-        Determine the new domain geometry with the absorbing layer.
-    check_timestep_abc()
-        Check if the timestep size is appropriate for the transient response.
-    create_mesh_with_layer()
-        Create a mesh with absorbing layer based on the determined size.
-    critical_boundary_points()
-        Determine critical boundary points using the Eikonal criterion.
     det_reference_freq()
         Determine the reference frequency for a new layer size.
-    geometry_infinite_model()
-        Determine the geometry for the infinite domain model.
-    identify_abc_layer_case()
-        Generate an identifier for the current layer geometry of the ABC.
-    infinite_model()
-        Create a reference model for the ABC scheme for comparative purposes.
-    layer_infinite_model()
-        Determine the domain extension size for the infinite domain model.
     layer_size_criterion()
         Determine the absorbing layer size using the Eikonal criterion for HABCs.
     nrbc_on_boundary_layer()
         Apply Non-Reflective BCs on the outer boundary of the absorbing layer.
-    path_to_save_abc_layer_case()
-        Create the path to save data for the current case study of the ABC scheme.
-    velocity_abc()
-        Set the velocity profile for the model with absorbing layer.
     """
 
     def __init__(self, domain_dim, frequency=None, dt=None,
@@ -215,30 +145,9 @@ class ABCLayer(NRBC, MeasureError):
         None
         """
 
-        # Model dimension
-        self.dimension = validate_parameter("dimension", dimension, [2, 3])
-
-        # Original domain dimensions
-        self.domain_dim = validate_data_structure("domain_dim", domain_dim, "tuple",
-                                                  expected_type_element=("float", "int"),
-                                                  expected_length=dimension)
-        # Source frequency
-        self.frequency = validate_numeric("frequency", frequency,
-                                          float_num=True, integer_num=True,
-                                          lower_bound=0., accept_parameter_as_none=True)
-
-        # Timestep for the transient simulation
-        self.dt = validate_numeric("dt", dt, float_num=True, integer_num=True,
-                                   lower_bound=0., accept_parameter_as_none=True)
-
-        # Nyquist frequency
-        self.freq_Nyquist = None if self.dt is None else 1. / (2. * self.dt)
-
-        # Quadrilateral/hexahedral elements
-        self.quadrilateral = quadrilateral
-
-        # Type of function space
-        self.func_space_type = func_space_type
+        AbsorbingBC.__init__(self, domain_dim, frequency=frequency, dt=dt,
+                             dimension=dimension, quadrilateral=quadrilateral,
+                             func_space_type=func_space_type, comm=comm)
 
         # ABC layer parameters
         self.abc_boundary_layer_type = validate_enum("abc_boundary_layer_type",
@@ -269,9 +178,6 @@ class ABCLayer(NRBC, MeasureError):
         # Define the shape of the absorbing layer.
         self.layer_geometry = self._define_layer_shape()
 
-        # Communicator MPI
-        self.comm = comm
-
         """"
         Create the path to save data
         The required abc_type argument from path_to_save_abc_layer_case() method is set to
@@ -285,8 +191,8 @@ class ABCLayer(NRBC, MeasureError):
                                   output_folder=output_folder)
 
         # Initializing the error measure class
-        MeasureError.__init__(self, output_folder=self.path_save,
-                              output_case=self.path_case_absl, comm=self.comm)
+        self.initialize_paths_for_error(output_folder=self.path_save,
+                                        output_case=self.path_case_absl)
 
     def _define_layer_shape(self):
         """Define the shape of the absorbing layer.
@@ -320,39 +226,6 @@ class ABCLayer(NRBC, MeasureError):
                                           n_type=self.abc_degree_type,
                                           dimension=self.dimension)
             return Hypershape_layer
-
-    def critical_boundary_points(self, wave):
-        """Determine critical boundary points using the Eikonal criterion.
-
-        Use original-domain boundaries to size the absorbing layer.
-        See Salas et al (2022): Hybrid absorbing scheme based on hyperelliptical
-        layers with non-reflecting boundary conditions in scalar wave equations.
-        doi: https://doi.org/10.1016/j.apm.2022.09.014
-        TODO: Add citation
-
-        Parameters
-        ----------
-        wave : `wave.Wave`
-            An instance of the :class:`~spyro.solvers.wave.Wave`.
-
-        Returns
-        -------
-        None
-        """
-
-        # Initializing Eikonal object
-        eik = Minimum_Eikonal(wave)
-
-        # Solving Eikonal
-        eik.solve_eik()
-
-        # Identifying critical points
-        self.eik_bnd = eik.ident_crit_eik()
-
-        # Critical point coordinates as receivers
-        pcrit = [tuple(bnd[0]) for bnd in self.eik_bnd]
-        wave.receiver_locations = pcrit + wave.receiver_locations
-        wave.number_of_receivers = len(wave.receiver_locations)
 
     def det_reference_freq(self, fpad=4):
         """Determine the reference frequency for a new layer size.
@@ -393,67 +266,6 @@ class ABCLayer(NRBC, MeasureError):
                 self.freq_ref = min(self.freq_ref, freq_ref)
 
         pprint(f"Reference Frequency (Hz): {self.freq_ref:.5f}", comm=self.comm)
-
-    def abc_new_geometry(self):
-        """Determine the new domain geometry with the absorbing layer.
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        -------
-        None
-        """
-
-        # Original domain dimensions
-        length_z, length_x = self.domain_dim[:2]
-
-        # New geometry with layer
-        self.length_xabc = length_x + 2 * self.abc_pad_length
-        self.length_zabc = length_z + self.abc_pad_length
-
-        if self.dimension == 3:  # 3D
-            length_y = self.domain_dim[2]
-            self.length_yabc = length_y + 2 * self.abc_pad_length
-
-    def abc_domain_dimensions(self, full_hyp=True):
-        """Return the new dimensions of the domain with absorbing layer as a tuple.
-
-        Parameters
-        ----------
-        full_hyp : `bool`, optional
-            Option to get the domain dimensions in hypershape layers.
-            If `True`, the domain dimensions with layer do not include truncation
-            due to the free surface. If `False`, the domain dimensions with layer
-            include truncation by free surface. Default is `True`.
-
-        Returns
-        -------
-        domain_layer : `tuple`
-            Domain dimensions with layer. For rectangular layers, truncation
-            due to the free surface is included (n = 1). For hypershape layers,
-            truncation by free surface is not included (n = 2) if 'full_hyp' is
-            `True`; otherwise, it is included (n = 1). (See Notes below)
-
-        Notes
-        -----
-        Model dimensions are defined as:
-            2D: (length_z + n * pad_len, length_x + 2 * pad_len).
-            3D: (length_z + n * pad_len, length_x + 2 * pad_len, length_y + 2 * pad_len).
-        """
-
-        # Domain dimensions with layer and truncations
-        domain_layer = [self.length_zabc, self.length_xabc]
-
-        # Domain dimensions with layer without truncations only for hypershape layers
-        if self.abc_boundary_layer_shape == LayerShapeType.HYPERSHAPE and full_hyp:
-            domain_layer[0] += self.abc_pad_length
-
-        if self.dimension == 3:  # 3D
-            domain_layer.append(self.length_yabc)
-
-        return tuple(domain_layer)
 
     def layer_size_criterion(self, lmin, fpad=4, n_root=1, layer_based_on_mesh=True):
         """Determine the absorbing layer size using the Eikonal criterion for HABCs.
@@ -527,168 +339,6 @@ class ABCLayer(NRBC, MeasureError):
             self.layer_geometry.calc_hyp_geom_prop(
                 domain_layer_full, self.abc_pad_length, self.lmin)
 
-    def create_mesh_with_layer(self, wave, inf_model=False, spln=True, save_file=True):
-        """Create a mesh with absorbing layer based on the determined size.
-
-        Parameters
-        ----------
-        wave : `wave.Wave`
-            An instance of the :class:`~spyro.solvers.wave.Wave`.
-        inf_model : `bool`, optional
-            If `True`, build a rectangular layer for the infinite or reference.
-            model (Model with "infinite" dimensions). Default is `False`.
-        spln : `bool`, optional
-            Flag to indicate whether to use splines (`True`) or lines (`False`).
-            in hypershape layer generation. Default is `True`.
-        save_file : `bool`, optional
-            If `True`, save the mesh with absorbing layer in a .pvd file.
-            Default is `True`.
-
-        Returns
-        -------
-        None
-        """
-
-        # Checking if the mesh for infinite model is requested
-        if inf_model:
-            pprint("\nGenerating Mesh for Infinite Model", comm=self.comm)
-            layer_shape = LayerShapeType.RECTANGULAR
-
-        else:
-            pprint("\nGenerating Mesh with Absorbing Layer", comm=self.comm)
-            layer_shape = self.abc_boundary_layer_shape
-
-        # New mesh with layer
-        if layer_shape == LayerShapeType.RECTANGULAR:
-
-            # Update the pad length in wave object
-            wave.abc_pad_length = self.abc_pad_length
-
-            # Create the mesh
-            wave.set_mesh()
-            pprint("Extended Rectangular Mesh Generated Successfully", comm=self.comm)
-
-        elif layer_shape == LayerShapeType.HYPERSHAPE:
-
-            # Update the pad length in wave.mesh_parameters object
-            wave.mesh_parameters.abc_pad_length = self.abc_pad_length
-
-            # Parameters for hypershape mesh
-            if self.dimension == 2:  # 2D
-                geometry_param = self.layer_geometry.perim_hyp
-
-            if self.dimension == 3:  # 3D
-                geometry_param = self.layer_geometry.surf_hyp
-
-            hypershape_param = (
-                self.layer_geometry.n_hyp, geometry_param, *self.layer_geometry.hyper_axes)
-
-            # Creating the mesh with the absorbing layer based on the hypershape geometry
-            mesh_abc = wave.mesh_ops.hypershape_mesh_habc(
-                hypershape_param, wave.mesh_original, wave.mesh_parameters, spln=spln)
-
-            # Updating the mesh with the absorbing layer
-            wave.set_mesh(user_mesh=mesh_abc)
-
-        pprint("Mesh Generated Successfully", comm=self.comm)
-
-        if save_file:
-            if inf_model:
-                pth_mesh = self.path_save + "preamble/mesh_inf.pvd"
-            else:
-                mesh_file_name = formatting_abc_layer_type("mesh_{}.pvd",
-                                                           self.abc_boundary_layer_type,
-                                                           for_prints=False)
-                pth_mesh = self.path_case_absl + mesh_file_name
-
-            # Save new mesh
-            outfile = VTKFile(pth_mesh)
-            outfile.write(wave.mesh)
-
-    def velocity_abc(self, wave, inf_model=False, method="point_cloud", save_file=True):
-        """Set the velocity profile for the model with absorbing layer.
-
-        Parameters
-        ----------
-        wave : `wave.Wave`
-            An instance of the :class:`~spyro.solvers.wave.Wave`.
-        inf_model : `bool`, optional
-            If `True`, build a rectangular layer for the infinite or reference
-            model (Model with "infinite" dimensions). Default is `False`.
-        method : `str`, optional
-            Method to extend the velocity profile. Options:
-            - "point_cloud" : Interpolate the field based on a point
-                              cloud from the original boundary
-            - "nearest_point" : Use the nearest point on the original
-                                boundary to extend the field.
-            Default is "point_cloud".
-        save_file : `bool`, optional
-            If `True`, save the velocity model with absorbing layer in a .pvd file.
-            Default is `True`.
-
-        Returns
-        -------
-        None
-
-        Notes
-        -----
-        "point_cloud" - dx = 0.05 km (2D)
-        Estimating Runtime and Used Memory
-        Runtime: (s):18.437, (m):0.307, (h):0.005
-        Used Memory: Current (MB):18.813, Peak (MB):25.102
-
-        "nearest_point" - dx = 0.05 km (2D)
-        Estimating Runtime and Used Memory
-        Runtime: (s):20.494, (m):0.342, (h):0.006
-        Used Memory: Current (MB):18.715, Peak (MB):25.298
-        """
-
-        pprint("\nUpdating Velocity Profile", comm=self.comm)
-
-        # Scalar space for auxiliar field of clipped coordinates
-        method_element = "DQ" if self.quadrilateral else "DG"
-        V = create_function_space(wave.mesh, method_element, 0)
-
-        # Initialize velocity field and assigning the original velocity model
-        wave.c = Function(V).interpolate(wave.initial_velocity_model,
-                                         allow_missing_dofs=True)
-
-        # Clipping coordinates to the layer domain
-        domain_layer = self.abc_domain_dimensions(full_hyp=False)
-        ufl_coordinates_abc = wave.mesh_ops.get_spatial_coordinates_abc(wave.mesh,
-                                                                        domain_layer)
-        lay_field, layer_mask = \
-            clipping_coordinates_lay_field(self.domain_dim, wave.mesh,
-                                           self.dimension, ufl_coordinates_abc,
-                                           V, quadrilateral=self.quadrilateral)
-
-        # Extending velocity model within the absorbing layer
-        extended_velocity = \
-            extend_scalar_field_profile(wave.mesh_original, wave.initial_velocity_model,
-                                        lay_field, layer_mask, wave.mesh_parameters.tol,
-                                        method=method, name_prop="Velocity")
-
-        # Interpolating the velocity model in the layer
-        wave.c.interpolate(extended_velocity * layer_mask + (1. - layer_mask)
-                           * wave.c, allow_missing_dofs=True)
-        del layer_mask, lay_field
-
-        # Interpolating in the space function of the problem
-        wave.c = Function(wave.function_space, name="c[km/s])").interpolate(wave.c)
-
-        # Save new velocity model
-        if save_file:
-            if inf_model:
-                pth_velocity = self.path_save + "preamble/c_inf.pvd"
-            else:
-                c_file_name = formatting_abc_layer_type("c_{}.pvd",
-                                                        self.abc_boundary_layer_type,
-                                                        for_prints=False)
-                pth_velocity = self.path_case_absl + c_file_name
-
-            outfile = VTKFile(pth_velocity)
-            outfile.write(wave.c)
-
     def nrbc_on_boundary_layer(self, wave, non_reflect_bc, save_file=True):
         """Apply Non-Reflective BCs on the outer boundary of the absorbing layer.
 
@@ -736,10 +386,9 @@ class ABCLayer(NRBC, MeasureError):
 
             if non_reflect_bc == BoundaryConditionsType.HIGDON:
                 crit_source = self.crit_source
-                bnd_nod_ids_nfs, bnd_nodes_nfs = \
-                    wave.mesh_ops.layer_boundary_data(wave.mesh,
-                                                      wave.function_space,
-                                                      wave.mesh_parameters)
+                bnd_nod_ids_nfs, bnd_nodes_nfs = wave.mesh_ops.layer_boundary_data(wave.mesh,
+                                                                                   wave.function_space,
+                                                                                   wave.mesh_parameters)
 
             # Hypershape parameters
             hyp_par = (self.layer_geometry.n_hyp, *self.layer_geometry.hyper_axes) \
@@ -751,238 +400,3 @@ class ABCLayer(NRBC, MeasureError):
 
         else:
             pprint("\nNot Non-Reflecting Boundary Conditions Prescribed", comm=self.comm)
-
-    def check_timestep_abc(self, wave, max_divisor_tf=1,
-                           set_max_dt=True, method='ANALYTICAL', mag_add=3):
-        """Check if the timestep size is appropriate for the transient response.
-
-        Parameters
-        ----------
-        wave : `acoustic_wave.AcousticWave`
-            An instance of the :class:`~spyro.solvers.acoustic_wave.AcousticWave`.
-        max_divisor_tf : `int`, optional
-            Index to select the maximum divisor of the final time, converted to an
-            integer according to the order of magnitude of the timestep size. The
-            timestep size is set to the divisor, given by the index in descending
-            order, less than or equal to the user's timestep size. If the value is 1,
-            the timestep size is set as the maximum divisor. Default is 1.
-        set_max_dt : `bool`, optional
-            If `True`, set the timestep size to the selected divisor. Default is `True`.
-        method : `str`, optional
-            Method to use for solving the eigenvalue problem. Default is 'ANALYTICAL'
-            method that estimates the maximum eigenvalue using the Gershgorin Circle
-            Theorem. Opts: 'ANALYTICAL', 'ARNOLDI', 'LANCZOS' or 'LOBPCG'.
-        mag_add : `int`, optional
-            Additional magnitude order to adjust the rounding of the timestep.
-
-        Returns
-        -------
-        None
-        """
-
-        # Cheking input parameters
-        validate_numeric("max_divisor_tf", max_divisor_tf,
-                         float_num=False, integer_num=True, lower_bound=0.)
-        validate_numeric("mag_add", mag_add, float_num=False, integer_num=True,
-                         lower_bound=0., include_lower_bound=True)
-
-        pprint("\nChecking Timestep Size", comm=self.comm)
-
-        # User timestep
-        usr_dt = wave.get_dt()
-        pprint(f"Current Nyquist Frequency: {self.freq_Nyquist:.5f} Hz", comm=self.comm)
-        pprint(f"Current Timestep Size: {1e3 * usr_dt:.{mag_add}f} ms", comm=self.comm)
-
-        # Maximum timestep size
-        dt_sol = Modal_Solver(self.dimension, method=method, calc_max_dt=True)
-        max_dt = dt_sol.estimate_timestep(wave.c, wave.function_space, wave.final_time,
-                                          shift=1e-8, quad_rule=wave.quadrature_rule,
-                                          fraction=1.)
-
-        # Rounding power
-        pot = int(abs(ceil(log10(max_dt))) + mag_add)
-
-        # Maximum timestep size according to divisors of the final time
-        val_int_tf = int(10**pot * wave.final_time)
-        val_int_dt = int(10**pot * max_dt)
-        max_div = [d for d in divisors(val_int_tf) if d < val_int_dt]
-        n_div = len(max_div)
-        index_div = min(max_divisor_tf, n_div)
-        max_dt = round(10**(-pot) * max_div[-index_div], pot)
-
-        # Set the timestep size
-        dt = max_dt if set_max_dt else min(usr_dt, max_dt)
-        wave.set_dt(dt)
-        dt_ms = 1e3 * wave.dt
-        if set_max_dt:
-            case_div = f"{min(max_divisor_tf, n_div)} of {n_div}"
-            str_dt = f"Selected Timestep Size ({case_div}): {dt_ms:.{mag_add}f} ms"
-        else:
-            str_dt = f"Selected Timestep Size: {dt_ms:.{mag_add}f} ms"
-
-        pprint(str_dt, comm=self.comm)
-
-        # Updating Nyquist frequency
-        self.freq_Nyquist = 1. / (2. * wave.dt)
-        pprint(f"New Nyquist Frequency: {self.freq_Nyquist:.5f} Hz", comm=self.comm)
-
-    def layer_infinite_model(self, lmin, c_bnd_max, final_time, source_locations=None):
-        """Determine the domain extension size for the infinite domain model.
-
-        Parameters
-        ----------
-        lmin : `float`
-            Minimum mesh size.
-        c_bnd_max : `float`
-            Maximum velocity value on the boundary of the original domain.
-        final_time : `float`
-            Final time of the simulation.
-        source_locations: `list`, optional
-            List of source locations.
-
-        Returns
-        -------
-        infinite_pad_len : `float`
-            Size of the domain extension for the infinite domain model.
-        """
-
-        # Cheking input parameters
-        validate_numeric("lmin", lmin, float_num=True,
-                         integer_num=True, lower_bound=0.)
-        validate_numeric("c_bnd_max", c_bnd_max, float_num=True,
-                         integer_num=True, lower_bound=0.)
-        validate_numeric("final_time", final_time, float_num=True,
-                         integer_num=True, lower_bound=0.)
-        validate_data_structure("source_locations", source_locations, "list",
-                                expected_type_element="tuple", accept_parameter_as_none=True)
-
-        # Size of the domain extension
-        add_dom = c_bnd_max * final_time / 2.
-
-        str_pad = "Infinite Domain Extension Based on "
-
-        # Distance already travelled by the wave
-        if hasattr(self, 'eik_bnd'):
-
-            str_pad += "Minimun Eikonal at Critical Boundary Points"
-
-            # If Eikonal analysis was performed (see `critical_boundary_points` method)
-            # Structure eikmin: [pnt_crit, c_bnd, eikmin, z_par, lref, sou_crit]
-            eikmin = self.eik_bnd[0][2]
-
-            # Minimum distance to the nearest boundary
-            dist_to_bnd = c_bnd_max * eikmin / 2.
-        else:
-
-            str_pad += "Minimum Distance Source-Boundary"
-
-            # If Eikonal analysis was not performed
-            source_loc = array(source_locations)
-
-            # Original  domain dimensions
-            length_z, length_x = self.domain_dim[:2]
-
-            # Candidate to minimum distance to the boundaries
-            delta_z = abs(source_loc[:, 0] - length_z)
-            delta_x = minimum(abs(source_loc[:, 1]), abs(source_loc[:, 1] - length_x))
-            cand_dist = (delta_z, delta_x)
-
-            if self.dimension == 3:  # 3D
-                length_y = self.domain_dim[2]
-                delta_y = minimum(abs(source_loc[:, 2]), abs(source_loc[:, 2] - length_y))
-                cand_dist += (delta_y,)
-
-            # Minimum distance to the nearest boundary
-            dist_to_bnd = float(min(cand_dist))
-
-        pprint(str_pad, comm=self.comm)
-
-        # Subtracting the distance already travelled by the wave
-        add_dom -= dist_to_bnd
-
-        # Pad length for the infinite domain extension
-        infinite_pad_len = lmin * ceil(add_dom / lmin)
-
-        return infinite_pad_len
-
-    def geometry_infinite_model(self, wave):
-        """Determine the geometry for the infinite domain model.
-
-        Parameters
-        ----------
-        wave : `acoustic_wave.AcousticWave`
-            An instance of the :class:`~spyro.solvers.acoustic_wave.AcousticWave`.
-
-        Returns
-        -------
-        None
-        """
-
-        lmin = wave.mesh_parameters.lmin if not hasattr(self, 'lmin') else self.lmin
-        c_bnd_max = wave.c_bnd_max
-        final_time = wave.final_time
-        source_locations = wave.source_locations
-
-        # Size of the domain extension
-        self.abc_pad_length = self.layer_infinite_model(lmin, c_bnd_max, final_time,
-                                                        source_locations=source_locations)
-        pprint(f"Infinite Domain Extension (km): {self.abc_pad_length:.4f}", comm=self.comm)
-
-        # New dimensions
-        self.abc_new_geometry()
-
-    def infinite_model(self, wave, check_dt=False, max_divisor_tf=1,
-                       method='ANALYTICAL', mag_add=3):
-        """Create a reference model for the ABC scheme for comparative purposes.
-
-        Parameters
-        ----------
-        wave : `acoustic_wave.AcousticWave`
-            An instance of the :class:`~spyro.solvers.acoustic_wave.AcousticWave`.
-        check_dt : `bool`, optional
-            If `True`, check if the timestep size is appropriate for the transient
-            response. Default is `False`.
-        max_divisor_tf : `int`, optional
-            Index to select the maximum divisor of the final time, converted to an
-            integer according to the order of magnitude of the timestep size. The
-            timestep size is set to the divisor, given by the index in descending
-            order, less than or equal to the user's timestep size. If the value is 1,
-            the timestep size is set as the maximum divisor. Default is 1.
-        method : `str`, optional
-            Method to use for solving the eigenvalue problem. Default is 'ANALYTICAL'
-            method that estimates the maximum eigenvalue using the Gershgorin Circle
-            Theorem. Opts: 'ANALYTICAL', 'ARNOLDI', 'LANCZOS' or 'LOBPCG'.
-        mag_add : `int`, optional
-            Additional magnitude order to adjust the rounding of the timestep.
-
-        Returns
-        -------
-        None
-        """
-
-        # Check the timestep size
-        if check_dt:
-            self.check_timestep_abc(wave, max_divisor_tf=max_divisor_tf,
-                                    method=method, mag_add=mag_add)
-
-        pprint("\nBuilding Infinite Domain Model", comm=self.comm)
-
-        # Defining geometry for infinite domain
-        self.geometry_infinite_model(wave)
-
-        # Creating mesh for infinite domain
-        self.create_mesh_with_layer(wave, inf_model=True)
-
-        # Updating velocity model
-        self.velocity_abc(wave, inf_model=True)
-
-        pprint("\nSolving Infinite Model", comm=self.comm)
-
-        # Solving the forward problem
-        wave.forward_solve()
-
-        # Saving reference signal
-        output_file = self.abc_boundary_layer_type.value + "_ref"
-        self.save_reference_signal(
-            wave.receiver_locations, wave.forward_solution_receivers,
-            wave.number_of_receivers, self.freq_Nyquist, output_file=output_file)
